@@ -186,11 +186,10 @@
     "@types/history": "4.7.8",
     "@types/jest": "^26.0.24",
     "@types/node": "15.0.2",
-    "@types/react": "17.0.2",
-    "@types/react-dom": "^17.0.9",
+    "@types/react": "^18.0.14",
+    "@types/react-dom": "^18.0.5",
     "@types/react-router": "5.1.18",
-    "@types/react-router-dom": "5.3.3",
-    "@types/react-test-renderer": "^17.0.1",
+    "@types/react-test-renderer": "^18.0.0",
     "@types/webpack-env": "^1.16.2",
     "@typescript-eslint/eslint-plugin": "^4.29.1",
     "@typescript-eslint/parser": "^4.29.1",
@@ -276,11 +275,11 @@
     "moment": "^2.29.2",
     "notistack": "^2.0.3",
     "nth-check": "2.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-router": "6.3.0",
+    "react-router-dom": "6.3.0",
+    "react-dom": "^18.2.0",
     "react-feather": "^2.0.9",
-    "react-router": "5.2.0",
-    "react-router-dom": "5.2.0",
     "regenerator-runtime": "^0.13.9",
     "source-map-support": "^0.5.19",
     "ts-proto": "^1.85.0",
