@@ -1,48 +1,44 @@
 import { ipcRenderer } from 'electron';
-import React, { FC, PropsWithChildren, useEffect } from 'react';
+import React, { FC, useEffect } from 'react';
 import {
-  useHistory,
   HashRouter,
-  Redirect,
-  Switch,
+  Outlet,
   Route,
+  Routes,
+  useNavigate,
+  Navigate,
 } from 'react-router-dom';
 
 import {
   CssBaseline,
   ThemeProvider,
   StyledEngineProvider,
-  Box,
 } from '@mui/material';
 import jssPreset from '@mui/styles/jssPreset';
 import StylesProvider from '@mui/styles/StylesProvider';
 import { create } from 'jss';
 import { SnackbarProvider } from 'notistack';
 import ConnectForm from './renderer/pages/ConnectForm';
-import TopBar from './renderer/components/TopBar';
 import ManageConnections from './renderer/pages/ManageConnections';
-import TopTabs from './renderer/components/TopTabs';
 import ConnectionView from './renderer/pages/ConnectionView';
 import SnackbarCloseButton from './renderer/components/SnackbarCloseButton';
 import { THEMES } from './shared/constants';
 import createCustomTheme, { ThemeConfig } from './shared/theme';
+import Layout from './renderer/pages/Layout';
 
-const RouteListener: FC = ({
-  children,
-}: PropsWithChildren<unknown>): JSX.Element => {
-  const history = useHistory();
+const RouteListener: FC = () => {
+  const navigate = useNavigate();
 
   useEffect(() => {
     ipcRenderer?.on('redirectTo', (_, arg) => {
-      history.replace(arg);
+      navigate(arg);
     });
     return function cleanup() {
       ipcRenderer.removeAllListeners('redirectTo');
     };
   }, []);
 
-  // eslint-disable-next-line react/destructuring-assignment
-  return <>{children}</>;
+  return <Outlet />;
 };
 
 const jss = create({ plugins: [...jssPreset().plugins] });
@@ -70,27 +66,26 @@ const App: FC = () => {
             )}
           >
             <HashRouter>
-              <TopBar>
-                <TopTabs />
-              </TopBar>
-              <Switch>
-                <RouteListener>
-                  <Route exact path="/">
-                    <Redirect to="/manage" />
+              <Routes>
+                <Route element={<RouteListener />}>
+                  <Route element={<Layout />}>
+                    <Route
+                      path="/"
+                      element={<Navigate to="/manage" replace />}
+                    />
+                    <Route path="/manage" element={<ManageConnections />} />
+                    <Route
+                      path="/view_connection/:connectionID"
+                      element={<ConnectionView />}
+                    />
+                    <Route path="/connectForm" element={<ConnectForm />} />
+                    <Route
+                      path="/edit_connect/:connectionID"
+                      element={<ConnectForm />}
+                    />
                   </Route>
-                  <Route exact path="/connectForm" component={ConnectForm} />
-                  <Route
-                    path="/edit_connect/:connectionID"
-                    component={ConnectForm}
-                  />
-                  <Route
-                    path="/view_connection/:connectionID"
-                    component={ConnectionView}
-                  />
-                  <Route exact path="/manage" component={ManageConnections} />
-                </RouteListener>
-              </Switch>
-              <Box mt={3} flexGrow={1} />
+                </Route>
+              </Routes>
             </HashRouter>
           </SnackbarProvider>
         </StylesProvider>
