@@ -1,6 +1,7 @@
 import React from 'react';
-import { render } from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import App from './App';
 import './App.global.css';
 
-render(<App />, document.getElementById('root'));
+const root = createRoot(document.getElementById('root') as HTMLDivElement);
+root.render(<App />);
